测试报文的发送

+++ userdefine/packagedefine.h
@@ -1,7 +1,7 @@
 #ifndef PACKAGEDEFINE_H
 #define PACKAGEDEFINE_H
 
-const short PACKAGE_SendField_ID = 0x01;
-const short PACKAGE_RecvField_ID = 0x01;
+const unsigned int  PACKAGE_SendField_ID = 0x01;
+const unsigned int  PACKAGE_RecvField_ID = 0x01;
 
 #endif
